RM: Added "pdfs" as a SPMS type to be uppercased.
In the future it is planned to remove the PS and PDF file types
from SPMS. The future type PDFS will replace them with "PDF or PS".
Since ICELAPCS 13 already started to use it and the central
repository is updated with the new file type PDFS I added it to
the JUDS scripts.

I kept the old types PS and PDF for compatibility with older SPMS
setups.